Google calendar integration

+++ src/main/java/com/evoapartments/accommodationbe/service/reservation/impl/ReservedAccommodationService.java
@@ -6,10 +6,14 @@ import com.evoapartments.accommodationbe.domain.accommodation.Accommodation;
 import com.evoapartments.accommodationbe.domain.reservation.ReservedAccommodation;
 import com.evoapartments.accommodationbe.repository.reservation.ReservedAccommodationRepository;
 import com.evoapartments.accommodationbe.service.accommodation.IAccommodationService;
+import com.evoapartments.accommodationbe.service.calendar.ICalendarSyncService;
+import com.evoapartments.accommodationbe.service.calendar.IGoogleCalendarService;
 import com.evoapartments.accommodationbe.service.reservation.IReservedAccommodationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
+import java.security.GeneralSecurityException;
 import java.util.List;
 
 @Service
@@ -17,6 +21,8 @@ import java.util.List;
 public class ReservedAccommodationService implements IReservedAccommodationService {
     private final ReservedAccommodationRepository accommodationReservedRepository;
     private final IAccommodationService accommodationService;
+    private final IGoogleCalendarService googleCalendarService;
+    private final ICalendarSyncService calendarSyncService;
 
     @Override
     public List<ReservedAccommodation> getAllReservations() {
@@ -28,54 +34,45 @@ public class ReservedAccommodationService implements IReservedAccommodationServi
     }
 
     @Override
-    public String saveReservation(Long accommodationId, ReservedAccommodation reservationRequest) {
-        if(reservationRequest.getCheckOutDate().isBefore(reservationRequest.getCheckInDate())){
-            throw  new InvalidReservationRequestException("Check-in date must be before Check-out date");
+    public String saveReservation(Long accommodationId, ReservedAccommodation request) throws IOException, GeneralSecurityException {
+        if (request.getCheckOutDate().isBefore(request.getCheckInDate())) {
+            throw new InvalidReservationRequestException("Check-in date must be before Check-out date");
         }
-        Accommodation accommodation = accommodationService.getAccommodationById(accommodationId).get();
-        List<ReservedAccommodation> existingReservations = accommodation.getReservations();
-
-        boolean accommodationIsAvailable = accommodationIsAvailable(reservationRequest, existingReservations);
-        if (accommodationIsAvailable){
-            accommodation.addReservation(reservationRequest);
-            accommodationReservedRepository.save(reservationRequest);
+        Accommodation accommodation = accommodationService.getAccommodationById(accommodationId).orElse(null);
+        if (accommodation != null) {
+            String calendarId = calendarSyncService.getGoogleCalendarIdExportByAccommodationId(accommodationId);
+            if (googleCalendarService.getAccommodationAvailabilityFromCalendarsFreeBusy(request.getCheckInDate().toString(),
+                    request.getCheckOutDate().toString(), accommodationId)) {
+                String eventId = googleCalendarService.scheduleReservationEventToGoogleCalendar(accommodationId, calendarId, request);
+                accommodation.addReservation(request, eventId);
+                accommodationReservedRepository.save(request);
+            } else {
+                throw new InvalidReservationRequestException("The accommodation has been booked for the selected dates.");
+            }
         } else {
-            throw new InvalidReservationRequestException("The accommodation has been booked for the selected dates.");
+            throw new InvalidReservationRequestException("The accommodation data fetch unsuccessful.");
         }
-        return reservationRequest.getReservationConfirmationCode();
+        return request.getReservationConfirmationCode();
     }
 
     @Override
     public ReservedAccommodation findByReservationConfirmationCode(String confirmationCode) {
         return  accommodationReservedRepository.findByReservationConfirmationCode(confirmationCode).orElseThrow(
-                () -> new ResourceNotFoundException("No reservation found with confirmation code: " + confirmationCode));
+                () -> new ResourceNotFoundException("There is no any reservation found with confirmation code: " + confirmationCode));
     }
 
     @Override
-    public void cancelReservation(Long reservationId) {
-        accommodationReservedRepository.deleteById(reservationId);
+    public void cancelReservation(Long reservationId) throws IOException {
+        ReservedAccommodation reservation = accommodationReservedRepository.findById(reservationId).orElse(null);
+        if (reservation != null) {
+            String calendarId = calendarSyncService.getGoogleCalendarIdExportByAccommodationId(reservation.getAccommodation().getId());
+            googleCalendarService.deleteGoogleCalendarReservationEvent(calendarId, reservation.getCalendarEventId());
+            accommodationReservedRepository.deleteById(reservationId);
+        }
     }
 
     @Override
     public List<ReservedAccommodation> getReservationsByUsername(String email) {
         return accommodationReservedRepository.findReservationsByGuestEmail(email);
-    }
-
-    private boolean accommodationIsAvailable(ReservedAccommodation reservationRequest, List<ReservedAccommodation> existingReservations) {
-        return existingReservations.stream()
-                .noneMatch(existingReservation ->
-                        reservationRequest.getCheckInDate().equals(existingReservation.getCheckInDate())
-                                || reservationRequest.getCheckOutDate().isBefore(existingReservation.getCheckOutDate())
-                                || (reservationRequest.getCheckInDate().isAfter(existingReservation.getCheckInDate())
-                                && reservationRequest.getCheckInDate().isBefore(existingReservation.getCheckOutDate()))
-                                || (reservationRequest.getCheckInDate().isBefore(existingReservation.getCheckInDate())
-                                && reservationRequest.getCheckOutDate().equals(existingReservation.getCheckOutDate()))
-                                || (reservationRequest.getCheckInDate().isBefore(existingReservation.getCheckInDate())
-                                && reservationRequest.getCheckOutDate().isAfter(existingReservation.getCheckOutDate()))
-                                || (reservationRequest.getCheckInDate().equals(existingReservation.getCheckOutDate())
-                                && reservationRequest.getCheckOutDate().equals(existingReservation.getCheckInDate()))
-                                || (reservationRequest.getCheckInDate().equals(existingReservation.getCheckOutDate())
-                                && reservationRequest.getCheckOutDate().equals(reservationRequest.getCheckInDate()))
-                );
     }
 }
